+++ v3/newrelic/internal_txn.go
@@ -413,17 +413,17 @@ func (thd *thread) End(recovered interface{}) error {
 			Category:     spanCategoryGeneric,
 			IsEntrypoint: true,
 		}
-		root.Attributes.addAttrs(txn.Attrs.Agent)
+		root.AgentAttributes.addAttrs(txn.Attrs.Agent)
 		if txn.rootSpanErrData != nil {
-			root.Attributes.addString(SpanAttributeErrorClass, txn.rootSpanErrData.Klass)
-			root.Attributes.addString(SpanAttributeErrorMessage, txn.rootSpanErrData.Msg)
+			root.AgentAttributes.addString(SpanAttributeErrorClass, txn.rootSpanErrData.Klass)
+			root.AgentAttributes.addString(SpanAttributeErrorMessage, txn.rootSpanErrData.Msg)
 		}
 		if nil != txn.BetterCAT.Inbound {
 			root.ParentID = txn.BetterCAT.Inbound.ID
 			root.TrustedParentID = txn.BetterCAT.Inbound.TrustedParentID
 			root.TracingVendors = txn.BetterCAT.Inbound.TracingVendors
 		}
-		root.Attributes = txn.Attrs.filterSpanAttributes(root.Attributes, destSpan)
+		root.AgentAttributes = txn.Attrs.filterSpanAttributes(root.AgentAttributes, destSpan)
 		txn.SpanEvents = append(txn.SpanEvents, root)
 
 		// Add transaction tracing fields to span events at the end of
@@ -1167,11 +1167,20 @@ func (txn *txn) Application() *Application {
 	return newApplication(txn.app)
 }
 
+// Note that Agent attributes added to spans must be on the allowed list of
+// span attributes, which you can find in attributes.go
 func (thd *thread) AddAgentSpanAttribute(key string, val string) {
 	txn := thd.txn
 	txn.Lock()
 	defer txn.Unlock()
 	thd.thread.AddAgentSpanAttribute(key, val)
+}
+
+func (thd *thread) AddUserSpanAttribute(key string, val interface{}) {
+	txn := thd.txn
+	txn.Lock()
+	defer txn.Unlock()
+	thd.thread.AddUserSpanAttribute(key, val)
 }
 
 var (